Remove requirement of no spaces before object operator ->
This commit removes the line from the ruleset requiring a space to
come before the arrow object operator ->. PSR2 does not enforce this
rule, and removing this rule will allow cleaner code when dealing
with fluent interfaces and method chaining. For example:

$obj->method1()
  ->method2()
  ->method3();

or

$obj
  ->method1()
  ->method2()
  ->method3();

This is especially useful when setting up PHPUnit mocks and their
constraints.